requirements.txt
discord.py>=2.3.0    # Or latest stable version of discord.py
deep-translator>=1.9.0 # Or latest stable version
python-dotenv>=0.20.0  # Or latest stable version